Maybe using float32 will fix Travis?

+++ tests/conftest.py
@@ -9,10 +9,10 @@ def data():
     (X_train, y_train), (X_test, y_test) = cifar10.load_data()
     y_train = keras.utils.to_categorical(y_train)
     y_test = keras.utils.to_categorical(y_test)
-    X_train = X_train.astype('float')
-    X_test = X_test.astype('float')
-    y_train = y_train.astype('float')
-    y_test = y_test.astype('float')
+    X_train = X_train.astype('float32')
+    X_test = X_test.astype('float32')
+    y_train = y_train.astype('float32')
+    y_test = y_test.astype('float32')
     X_train /= 255
     X_test /= 255
     return (X_train, y_train), (X_test, y_test)
